# backend/APIDocumentation.yaml
openapi: 3.0.0
info:
  title: Netro API
  description: |
    API for the Netro social platform. This API handles user authentication, profile management, 
    categories, email verification, and events.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Response Format
    All responses follow a consistent format:
    - **Success responses** include `success: true` and a `message` field
    - **Error responses** include `success: false` and a descriptive `message` field
    - Validation errors include an additional `errors` array
    - Invalid category errors include an `invalidCategories` array
  version: 1.0.0
  contact:
    name: Netro API Support
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /event/create:
    post:
      tags:
        - Events
      summary: Create a new event
      description: Create a new event with title, description, location, and optional categories. If seriesEvent is false, frequency is automatically set to null.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - startTime
                - location
              properties:
                title:
                  type: string
                  maxLength: 100
                  description: Event title
                  example: Basketball Game
                description:
                  type: string
                  maxLength: 255
                  description: Event description
                  example: Friendly basketball match at the park
                startTime:
                  type: string
                  format: date-time
                  description: Event start time in ISO 8601 format
                  example: '2026-03-15T18:00:00Z'
                location:
                  type: string
                  maxLength: 100
                  description: Event location
                  example: Central Park
                seriesEvent:
                  type: boolean
                  default: false
                  description: Whether this is a recurring event
                  example: true
                frequency:
                  type: string
                  maxLength: 100
                  nullable: true
                  description: Recurrence frequency (e.g., 'weekly', 'monthly'). Automatically set to null if seriesEvent is false.
                  example: weekly
                categories:
                  type: array
                  items:
                    type: string
                  description: Array of category names associated with the event
                  example: ["basketball", "gaming"]
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCreateSuccessResponse'
        '400':
          description: Validation failed or invalid categories
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/InvalidCategoriesError'
              examples:
                missingFields:
                  value:
                    success: false
                    message: Validation failed
                    errors: ["Title is required", "Description is required", "Start time is required", "Location is required"]
                invalidStartTime:
                  value:
                    success: false
                    message: Validation failed
                    errors: ["Invalid start time format"]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: User not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Failed to create event
  /event/edit:
    put:
      tags:
        - Events
      summary: Update an existing event
      description: Update an existing event. Only the event creator can edit. If seriesEvent is set to false, frequency is automatically set to null. At least one field besides eventId must be provided.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
              properties:
                eventId:
                  type: integer
                  description: ID of the event to update
                  example: 1
                title:
                  type: string
                  maxLength: 100
                  description: Updated event title
                  example: Updated Basketball Game
                description:
                  type: string
                  maxLength: 255
                  description: Updated event description
                  example: Updated description for the basketball match
                startTime:
                  type: string
                  format: date-time
                  description: Updated event start time
                  example: '2026-03-16T19:00:00Z'
                location:
                  type: string
                  maxLength: 100
                  description: Updated event location
                  example: Downtown Court
                seriesEvent:
                  type: boolean
                  description: Whether this is a recurring event. Setting to false automatically sets frequency to null.
                  example: false
                frequency:
                  type: string
                  maxLength: 100
                  nullable: true
                  description: Updated recurrence frequency
                  example: monthly
                categories:
                  type: array
                  items:
                    type: string
                  description: Updated array of category names. Replaces all existing categories.
                  example: ["basketball", "football"]
              minProperties: 2
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventUpdateSuccessResponse'
        '400':
          description: Validation failed or no fields provided
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/InvalidCategoriesError'
                  - $ref: '#/components/schemas/Error'
              examples:
                noFields:
                  value:
                    success: false
                    message: At least one field must be provided to update
                noEventId:
                  value:
                    success: false
                    message: Validation failed
                    errors: ["Event ID is required"]
                invalidCategories:
                  value:
                    success: false
                    message: Invalid categories provided
                    invalidCategories: ["invalid_category"]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - user does not own this event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: You do not have permission to edit this event
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Event not found
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                message: Failed to update event
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
    Error:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
    ValidationError:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: string
          example: ["Email is required", "Password must be at least 8 characters"]
    InvalidCategoriesError:
      type: object
      required:
        - success
        - message
        - invalidCategories
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid categories provided
        invalidCategories:
          type: array
          items:
            type: string
          example: ["invalid_category_1", "invalid_category_2"]

    # User schemas
...
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Basketball Game
        description:
          type: string
          example: Friendly basketball match at the park
        startTime:
          type: string
          format: date-time
          example: '2026-03-15T18:00:00.000Z'
        location:
          type: string
          example: Central Park
        seriesEvent:
          type: boolean
          example: true
        frequency:
          type: string
          nullable: true
          example: weekly
        createdAt:
          type: string
          format: date-time
          example: '2024-02-13T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-02-13T10:30:00.000Z'
        createdByUserId:
          type: integer
          example: 1
        categories:
          type: array
          items:
            type: string
          example: ["basketball", "gaming"]

    # Success response schemas
...
    EventCreateSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            event:
              $ref: '#/components/schemas/Event'
    EventUpdateSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            event:
              $ref: '#/components/schemas/Event'